serialization fix

+++ trace-cobertura/src/main/java/se/de/hu_berlin/informatik/spectra/provider/tracecobertura/infrastructure/BufferedArrayQueue.java
@@ -115,7 +115,7 @@ public class BufferedArrayQueue<E> extends AbstractQueue<E> implements Serializa
         currentStoreIndex = stream.readInt();
         lastStoreIndex = stream.readInt();
         firstNodeSize = stream.readInt();
-        size = stream.readInt();
+        size = stream.readLong();
         arrayLength = stream.readInt();
         
         cachedNodes = new HashMap<>();
