+++ trace-cobertura/src/main/java/se/de/hu_berlin/informatik/spectra/provider/tracecobertura/infrastructure/BufferedLongArrayQueue.java
@@ -108,7 +108,7 @@ public class BufferedLongArrayQueue implements Serializable {
         currentStoreIndex = stream.readInt();
         lastStoreIndex = stream.readInt();
         firstNodeSize = stream.readInt();
-        size = stream.readInt();
+        size = stream.readLong();
         arrayLength = stream.readInt();
         
         cachedNodes = new HashMap<>();
